image_proccessing_service/requirements.txt
certifi==2022.12.7
cffi==1.15.1
minio==7.1.13
pika==1.3.1
Pillow==9.4.0
pycparser==2.21
pyheif==0.7.1
urllib3==1.26.14